pyproject.toml
[project]
name = "mcp-local-reference"
version = "0.1.0"
description = "MCP server for accessing Zotero references, extracting PDF text, and cropping figures"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
keywords = ["mcp", "zotero", "references", "pdf", "figures", "academic"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "mcp[cli]>=1.0.0",
    "pymupdf>=1.25.0",
    "pillow>=10.0.0",
    "chromadb>=0.6.0",
]

[project.scripts]
mcp-local-reference = "mcp_local_reference.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_local_reference"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
packages = ["mcp_local_reference"]
mypy_path = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
